equiv: tier-7 writable kmem, timers, helper tail — and an object-hygiene fix
Model:
- kmem is WRITABLE and OBSERVABLE: a store through a data-valued pointer
  (a socket/task pointer from the ctx, a probe-read'd address) writes the
  same shared array probe-reads read, replacing a blanket bail (34
  objects). Control: mutating the value stored through ctx->sk makes the
  kmem observable alone differ.
- Timers: init/start/cancel are trace events + errno oracle;
  set_callback records the slot, its map and the callback's SOURCE SYMBOL
  NAME. Executing the callback inline at registration was tried and
  REMOVED: it interleaves the callback's own events at a point they never
  occur and flagged equivalent programs. The callback BODY is therefore
  not proved — documented as a deliberate gap.
- Re-typing casts (skc_to_*, bpf_tcp_sock) alias their argument rather
  than returning a fresh region (a fresh region made reads through the
  cast disagree with reads through the original).
- Helper tail: get_func_arg/ret, get_netns_cookie, strncmp,
  skb_get/set_tunnel_opt, per_cpu/this_cpu_ptr, kptr_xchg, and
  side-effect entries for clone_redirect, redirect_peer,
  skb_change_proto/type/tail, sk_select_reuseport (key compared by
  CONTENTS — the compilers place it at different stack offsets).
- normalize_name: multi-digit v0 length prefixes never demangled, so any
  Rust symbol of 10+ characters stayed mangled.

OBJECT HYGIENE (important): `make test-<name>` swaps the Rust object into
the selftests slot and only `make restore-<name>` undoes it; an
interrupted run leaves later comparisons Rust-vs-Rust and trivially
"equivalent". An audit found 20 contaminated objects. check.py now
refuses to run when a C object differs from its .corig, `make
restore-all` fixes the tree, and restoring exposed two divergences the
contamination had masked (test_sockmap_listen, test_tunnel_kern).

Corpus: 1158 EQUIV / 3 INEQUIV / 16 WAIVED (307/550 objects, was 268).
True findings #17-#25, all fixed and QEMU-gated: 13 dropped bpf_printk
sites (timer, test_tunnel_kern — incl. 4-/5-arg prints that lower to
trace_vprintk/177); timer_crash's `== 1` where C uses truthiness, plus
struct padding leaking into a map value; bpf_iter_setsockopt's bool and a
zero-init where C leaves the buffer uninitialized; timer_start_deadlock
masking a bool ctx arg to a byte; tracing_struct staging get_func_arg
through zeroed locals; test_tc_dtime masking skb->protocol;
test_sockmap_listen's bools — where clang emitted `jne 1` at three sites
and `jne 0` at a fourth IN THE SAME FILE. The erspan divergences are a
bug in the C selftest (BPF_CORE_WRITE_BITFIELD roots its relo at
erspan_metadata, not erspan_md2, so md2 writes land past the struct) —
waived with that diagnosis. Open: 3 test_tc_dtime programs.

Also reverts an llvm.memset -> bpf_arena_memset rewrite that leaked from
the collections pipeline into rust-bpf/add_ksyms.py: it made rebuilt
objects reference a kfunc the kernel has no BTF for, so they failed to
load (timer went 7 FAILED -> 14/14 PASSED after the revert).

+++ progs/test_tc_dtime.rs
@@ -281,8 +281,10 @@ fn skb_get_type(skb: *const __sk_buff) -> i32 {
     let mut ns: u8 = 0;
     let trans: usize;
 
-    let protocol = vload!((*skb).protocol) as u16;
-    if protocol == htons(ETH_P_IP) {
+    // C's `switch (skb->protocol)` compares the FULL 32-bit ctx word; a
+    // 16-bit mask here diverges for words whose upper bits are set.
+    let protocol = vload!((*skb).protocol) as u32;
+    if protocol == htons(ETH_P_IP) as u32 {
         let iph = (data + ETH_HDR_LEN) as *const iphdr;
         if iph as usize + core::mem::size_of::<iphdr>() > data_end {
             return -1;
@@ -295,7 +297,7 @@ fn skb_get_type(skb: *const __sk_buff) -> i32 {
         }
         inet_proto = unsafe { (*iph).protocol };
         trans = iph as usize + core::mem::size_of::<iphdr>();
-    } else if protocol == htons(ETH_P_IPV6) {
+    } else if protocol == htons(ETH_P_IPV6) as u32 {
         let ip6h = (data + ETH_HDR_LEN) as *const ipv6hdr;
         if ip6h as usize + core::mem::size_of::<ipv6hdr>() > data_end {
             return -1;
